==> contrib/derecho/step4_convert_to_scm_input.derecho.ipynb <==
%matplotlib inline
import xarray as xr
import numpy as np
import metpy.calc as mpcalc
from metpy.units import units
import matplotlib.pyplot as plt
import pandas as pd

def transpose_da(da, dim_list):
    if len(dim_list) == 1:
        return da
    if len(dim_list) == 2:
        return da.transpose(dim_list[0],dim_list[1])
    if len(dim_list) == 3:
        return da.transpose(dim_list[0],dim_list[1],dim_list[2])
    if len(dim_list) == 4:
        return da.transpose(dim_list[0],dim_list[1],dim_list[2],dim_list[3])
    if len(dim_list) == 5:
        return da.transpose(dim_list[0],dim_list[1],dim_list[2],dim_list[3],dim_list[4])

def format_da_to_target(da_replace, da_temp_var, data_type):
    if data_type == 'forcing':
        dim_replace = []
        var_coords_dic = {}
        for dim in da_temp_var.dims:
            dim_replace.append(coord_name_dic_forcing[dim])
            if dim in coords_dic.keys():
                var_coords_dic.update({dim:coords_dic[dim]})
        da_replace_t = transpose_da(da_replace,dim_replace)
        da_replace_n = xr.DataArray(da_replace_t.values, dims=da_temp_var.dims,
                                    coords = var_coords_dic)
        da_replace_n.attrs = da_temp_var.attrs
    if data_type == 'scalar':
        if len(np.shape(da_replace.values)) == 0:
            value = np.float64(da_replace.values)
        else:
            value = np.float64(da_replace.values[0])
        
        da_replace_n = xr.DataArray(value)
        da_replace_n.attrs = da_temp_var.attrs
    if data_type == 'initial':
        dims_tuple = ('t0',)
        var_coords_dic = {'t0':coords_dic['t0']}
        da_init = da_replace.isel(time=0)
        if len(da_init.dims)>0:
            for dim in da_init.dims:
                if dim in coord_name_dic_reverse.keys():
                    dim_target = coord_name_dic_reverse[dim]
                    dims_tuple += (dim_target,)
                    if dim_target in coords_dic.keys():
                        var_coords_dic.update({dim:coords_dic[dim_target]})
        da_replace_n = xr.DataArray([da_init.values], dims=dims_tuple,
                                  coords = var_coords_dic)
        da_replace_n.attrs = da_temp_var.attrs

    return da_replace_n

temp_file = "/glade/work/yifanc/code/scm_v7/ccpp-scm/scm/data/processed_case_input/gabls3_noahmp_SCM_driver.nc"
static_file = "/glade/work/yifanc/proj/wpo2023/static/ameriflux/ameriflux_static_fields.C1152.US-Whs.nc"
forcing_file = "/glade/derecho/scratch/yifanc/wpo2023/SCM/processed/US-Whs/processed_era5.201902.US-Whs.derecho.nc"

excel_file = "ERA5_variable.SCM.input.xlsx"
excel_file = pd.ExcelFile(excel_file)
mapping_df = excel_file.parse(sheet_name="var_mapping")

mapping_df.set_index("scm_var_name", inplace=True)

ds_temp = xr.open_dataset(temp_file)
ds_static = xr.open_dataset(static_file)
ds_forcing = xr.open_dataset(forcing_file)

time_dims = ds_forcing["time"]
t0 = time_dims[0].values

coord_list = []
for coord in ds_temp.coords:
    print(coord)
    coord_list.append(coord)

out = xr.Dataset()

lev_coords = xr.DataArray(ds_forcing.levels.values,dims=['lev'],
                          coords={'lev':ds_forcing.levels.values})
lev_coords.attrs = ds_temp.lev.attrs

t0_coords = xr.DataArray(ds_forcing.time.values[[0]],dims=['t0'],
                         coords={'t0':ds_forcing.time.values[[0]]})
t0_coords.attrs = ds_temp.t0.attrs
time_coords = xr.DataArray(ds_forcing.time.values,dims=['time'],
                         coords={'time':ds_forcing.time.values})
time_coords.attrs = ds_temp.time.attrs


out.coords['lev'] = lev_coords
out.coords['t0'] = t0_coords
out.coords['time'] = time_coords

coords_dic = {'lev':lev_coords,
              't0':t0_coords,
              'time':time_coords}

data_dic = {'forcing':ds_forcing,
            'static':ds_static}

coord_name_dic_forcing = {'lev':'levels',
                         'lat':'latitude',
                         'lon':'longitude',
                         'time':'time',
                         't0':'t0',
                         'nsoil':'nsoil'}
coord_name_dic_reverse = {'levels':'lev',
                         'latitude':'lat',
                         'longitude':'lon',
                         'time':'time',
                         't0':'t0',
                         'nsoil':'nsoil'}

nudging_var_list = ['ua','va','ta','qv']
nudging_invar_dic = {'ta':'T_nudge',
                     'qv':'qt_nudge',
                     'ua':'u_nudge',
                     'va':'v_nudge'}
nudging_res_lev = 50000
nudging_timestep = 2 * 3600 

for var in ds_temp.variables:
    if var not in coord_list:
        data_source = mapping_df.loc[var, "data_source"]
        vname_source = mapping_df.loc[var, "data_source_var_name"]
        data_type = mapping_df.loc[var, "data_type"]
        print(var,data_type)
        da_temp_var = ds_temp[var]
        if isinstance(data_source, str):
            print(data_source,var, vname_source, data_type)
            ds_source = data_dic[data_source]
            da_replace = ds_source[vname_source]
            da_replace_n = format_da_to_target(da_replace,da_temp_var,data_type)
            out[var] = da_replace_n
        elif 't0' in da_temp_var.dims:
            if 'lev' in da_temp_var.dims:
                da_temp_var = xr.DataArray(np.full([1,len(lev_coords)],0),
                                           dims=['t0','lev'],
                                           coords={'t0':coords_dic['t0'],
                                                   'lev':coords_dic['lev']})
                out[var] = da_temp_var
        else:
            da_temp_var_no_change = ds_temp[var]
            out[var] = da_temp_var_no_change

out.attrs = ds_temp.attrs

out.attrs['start_date'] = f"{str(time_coords.values[0])[0:10]} {str(time_coords.values[0])[11:19]}"
out.attrs['end_date'] = f"{str(time_coords.values[-1])[0:10]} {str(time_coords.values[-1])[11:19]}"

out_lev = out.sortby('lev', ascending=False)

# era5_noahmp_exp_SCM_driver.test10.nc
out_lev = out_lev.rename(name_dict={"thetal":"theta",
                          "tnthetal_adv":"tntheta_adv"})
out_lev.attrs['adv_theta'] = np.int32(1)
out_lev.attrs['adv_thetal'] = np.int32(0)

# era5_noahmp_exp_SCM_driver.test15.nc
da_temp_var = ds_temp['pa_forc']
data_type = 'forcing'
for nvar in nudging_var_list:
    invar = nudging_invar_dic[nvar]
    da = ds_forcing[invar]
    da_nudging = format_da_to_target(da,da_temp_var,data_type)
    da_nudging.attrs = da.attrs
    out_lev[f"{nvar}_nud"] = da_nudging
    out_lev.attrs[f"pa_nudging_{nvar}"] = nudging_res_lev
    out_lev.attrs[f"nudging_{nvar}"] = nudging_timestep

out_lev.to_netcdf("/glade/derecho/scratch/yifanc/wpo2023/SCM/processed/US-Whs/era5_noahmp_exp_SCM_driver.nc")
